ci: Makes 'latest' CI jobs fail on error
Updates 'latest' Snakemake job configurations (Singularity & Apptainer) to fail the CI pipeline on error by setting `allow_failure` to `false`. This change reflects that these configurations are now considered stable.

Adds comments to explain the necessity of the `--bind` argument for the `make_linked_txome` rule when running on GitHub Actions with containerization.